add a plugin system
this commit adds a plugin system to the bot.

the plugin system is based on the idea that each plugin is a node module
that exports a single function.

the function is called with the bot object as the first argument and an
options object as the second argument.

the options object is passed to the plugin by the user in the config file.

the plugin can then use the bot object to register listeners and commands.

the plugin can also use the options object to configure itself.